Notes

+++ src/main/java/com/example/demo/services/impl/NoteServiceImpl.java
@@ -3,10 +3,11 @@ package com.example.demo.services.impl;
 import com.example.demo.models.Note;
 import com.example.demo.models.User;
 import com.example.demo.repositories.NoteRepository;
-import com.example.demo.repositories.UserRepository;
 import com.example.demo.services.NoteService;
+import com.example.demo.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
@@ -25,7 +26,7 @@ public class NoteServiceImpl implements NoteService {
     }
 
     @Autowired
-    private UserRepository userRepository;
+    private UserService userService;
 
     @Override
     public List<Note> getAll() {
@@ -39,15 +40,12 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public Note addNoteToUserById(Integer id, String text) {
-
-        User user = userRepository.findOne(id);
-
+        User user = userService.findOne(id);
         if (null == user) {
             throw new EntityNotFoundException(String.format("There is no user with such id: %d", id));
         }
 
         Note newNote = new Note();
-
         newNote.setText(text);
         newNote.setUser(user);
 
@@ -56,9 +54,7 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public List<Note> getAllUserNotesById(Integer id) {
-
-        User user = userRepository.findOne(id);
-
+        User user = userService.findOne(id);
         if (null == user) {
             throw new EntityNotFoundException(String.format("There is no user with such id: %d", id));
         }
@@ -67,28 +63,44 @@ public class NoteServiceImpl implements NoteService {
     }
 
     @Override
-    public void removeNoteById(Integer noteId) {
+    public void removeNoteById(Integer userId, Integer noteId) {
         Note note = noteRepository.getOne(noteId);
-
         if (null == note) {
             throw new EntityNotFoundException(String.format("There is no note with such id: %d", noteId));
+        }
+
+        if (!note.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException(String.format("Note doesn't belong to the user with id: %d", userId));
         }
 
         noteRepository.delete(note);
     }
 
     @Override
+    @Transactional
     public void removeAllUserNotes(Integer id) {
-
-        User user = userRepository.findOne(id);
-
+        User user = userService.findOne(id);
         if (null == user) {
             throw new EntityNotFoundException(String.format("There is no user with such id: %d", id));
         }
 
         List<Note> notes = noteRepository.findByUserId(id);
-
         noteRepository.delete(notes);
+    }
+
+    @Override
+    public Note updateNoteById(Integer userId, Integer noteId, String noteText) {
+        Note note = noteRepository.getOne(noteId);
+        if (null == note) {
+            throw new EntityNotFoundException(String.format("There is no note with such id: %d", noteId));
+        }
+
+        if (!note.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException(String.format("Note doesn't belong to the user with id: %d", userId));
+        }
+
+        note.setText(noteText);
+        return noteRepository.saveAndFlush(note);
     }
